need to fetch crossfilter, dimension through accessors
because they can and will change

+++ src/delete_things.js
@@ -1,12 +1,18 @@
-dc_graph.delete_things = function(crossfilter, dimension, things_group, mode_name) {
+dc_graph.delete_things = function(things_group, mode_name) {
     var _keyboard, _selected = [];
     function selection_changed(selection) {
         _selected = selection;
     }
     function delete_selection() {
+        if(!_behavior.crossfilterAccessor())
+            throw new Error('need crossfilterAccessor');
+        if(!_behavior.dimensionAccessor())
+            throw new Error('need dimensionAccessor');
         var authorize = _behavior.onDelete() ? _behavior.onDelete()(_selected) : Promise.resolve(_selected);
         authorize.then(function(selection) {
             if(selection && selection.length) {
+                var crossfilter = _behavior.crossfilterAccessor()(_behavior.parent()),
+                    dimension = _behavior.dimensionAccessor()(_behavior.parent());
                 dimension.filterFunction(function(k) {
                     return selection.indexOf(k) !== -1;
                 });
@@ -37,5 +43,7 @@ dc_graph.delete_things = function(crossfilter, dimension, things_group, mode_nam
         }
     });
     _behavior.onDelete = property(null);
+    _behavior.crossfilterAccessor = property(null);
+    _behavior.dimensionAccessor = property(null);
     return _behavior;
 };
